requirements.txt
mediapipe
streamlit
numpy
opencv_python_headless
pillow